refactor(clawdstrike-logos): extract inheritance from verifier.rs
Move the semantic inheritance soundness engine
(inspect_policy_inheritance_against_parent, the per-guard check_*_inheritance
fns, disabled_*_config builders, the MCP max-args probe, and domain_action)
into verifier/inheritance.rs. Pure relocation: free fns widened to pub(crate)
for cross-module resolution via the root glob; the file-local
DEFAULT_MCP_MAX_ARGS_SIZE const stays private. No logic changes.
